Move command handlers to new file
